+++ packages/salesforcedx-vscode-core/src/commands/forceSourceDeployManifest.ts
@@ -4,21 +4,27 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { ContinueResponse, SfdxCommandBuilder } from '@salesforce/salesforcedx-utils-vscode';
+import {
+  ContinueResponse,
+  SfdxCommandBuilder
+} from '@salesforce/salesforcedx-utils-vscode';
 import { ComponentSet } from '@salesforce/source-deploy-retrieve';
 import { join } from 'path';
 import * as vscode from 'vscode';
 import { channelService } from '../channels';
-import {
-  TimestampConflictChecker
-} from '../commands/util/postconditionCheckers';
+import { TimestampConflictChecker } from '../commands/util/postconditionCheckers';
 import { nls } from '../messages';
 import { notificationService } from '../notifications';
 import { SfdxPackageDirectories } from '../sfdxProject';
 import { telemetryService } from '../telemetry';
 import { workspaceUtils } from '../util';
-import { DeployExecutor } from './baseDeployRetrieve';
-import { ConflictDetectionMessages, FilePathGatherer, SfdxCommandlet, SfdxWorkspaceChecker } from './util';
+import { DeployExecutor } from './deployExecutor';
+import {
+  ConflictDetectionMessages,
+  FilePathGatherer,
+  SfdxCommandlet,
+  SfdxWorkspaceChecker
+} from './util';
 
 export class LibrarySourceDeployManifestExecutor extends DeployExecutor<
   string
