+++ packages/salesforcedx-vscode-core/src/commands/forceSourceRetrieveManifest.ts
@@ -19,7 +19,7 @@ import { sfdxCoreSettings } from '../settings';
 import { SfdxPackageDirectories } from '../sfdxProject';
 import { telemetryService } from '../telemetry';
 import { workspaceUtils } from '../util';
-import { RetrieveExecutor } from './baseDeployRetrieve';
+import { RetrieveExecutor } from './deployRetrieveExecutor';
 import {
   FilePathGatherer,
   SfdxCommandlet,
