+++ packages/salesforcedx-vscode-core/src/conflict/metadataCacheService.ts
@@ -20,7 +20,7 @@ import * as os from 'os';
 import * as path from 'path';
 import * as shell from 'shelljs';
 import * as vscode from 'vscode';
-import { RetrieveExecutor } from '../commands/baseDeployRetrieve';
+import { RetrieveExecutor } from '../commands/deployRetrieveExecutor';
 import { WorkspaceContext } from '../context/workspaceContext';
 import { setApiVersionOn } from '../services/sdr/componentSetUtils';
 import { SfdxPackageDirectories } from '../sfdxProject';
